Move Hashable type-trait to own subproject.

+++ pico-hash/src/main/scala/org/pico/hash/Hashable.scala
@@ -1,5 +1,5 @@
 package org.pico.hash
 
 trait Hashable[A] {
-  def hash(a: A): Int
+  def hash(a: A): Long
 }
